added configurable maps logo

+++ src/Boxtal/BoxtalConnectWoocommerce/util/class-configuration-util.php
@@ -37,6 +37,26 @@ class Configuration_Util {
 		return $url . '?' . http_build_query( $params );
 	}
 
+    /**
+     * Get map logo href url.
+     *
+     * @return string map logo href url
+     */
+    public static function get_map_logo_href_url() {
+        $url = get_option('BW_MAP_LOGO_HREF_URL');
+        return false !== $url ? $url : null;
+    }
+
+    /**
+     * Get map logo image url.
+     *
+     * @return string map logo image url
+     */
+    public static function get_map_logo_image_url() {
+        $url = get_option('BW_MAP_LOGO_IMAGE_URL');
+        return false !== $url ? $url : null;
+    }
+
 	/**
 	 * Has configuration.
 	 *
@@ -58,6 +78,8 @@ class Configuration_Util {
 		delete_option( 'BW_SECRET_KEY' );
 		delete_option( 'BW_MAP_BOOTSTRAP_URL' );
 		delete_option( 'BW_MAP_TOKEN_URL' );
+		delete_option( 'BW_MAP_LOGO_IMAGE_URL' );
+		delete_option( 'BW_MAP_LOGO_HREF_URL' );
 		delete_option( 'BW_PP_OPERATORS' );
 		delete_option( 'BW_TRACKING_EVENTS' );
 		delete_option( 'BW_NOTICES' );
@@ -147,11 +169,16 @@ class Configuration_Util {
 	 * @return boolean
 	 */
 	private static function parse_map_configuration( $body ) {
-		if ( is_object( $body ) && property_exists( $body, 'mapsBootstrapUrl' ) && property_exists( $body, 'mapsTokenUrl' ) ) {
+		if ( is_object( $body ) && property_exists( $body, 'mapsBootstrapUrl' ) && property_exists( $body, 'mapsTokenUrl' )
+            && property_exists( $body, 'mapsLogoImageUrl' ) && property_exists( $body, 'mapsLogoHrefUrl' )) {
             //phpcs:ignore
             update_option('BW_MAP_BOOTSTRAP_URL', $body->mapsBootstrapUrl);
             //phpcs:ignore
             update_option('BW_MAP_TOKEN_URL', $body->mapsTokenUrl);
+            //phpcs:ignore
+            update_option('BW_MAP_LOGO_IMAGE_URL', $body->mapsLogoImageUrl);
+            //phpcs:ignore
+            update_option('BW_MAP_LOGO_HREF_URL', $body->mapsLogoHrefUrl);
 			return true;
 		}
 		return false;
